backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.2
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
httpx==0.27.2
tenacity==9.0.0
celery==5.4.0
redis==5.0.8
boto3==1.35.20
moviepy==1.0.3
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
loguru==0.7.2
orjson==3.10.7
aiofiles==24.1.0
apscheduler==3.10.4
prometheus-client==0.21.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.145.0
psutil==6.0.0
